requirements.txt
streamlit
pandas
plotly
newsapi-python
nltk
yfinance